refactor: centralise homepage banners into dedicated Banners component
