--- src/components/CurriculumComponents/Banner/index.tsx
@@ -1,1 +1,0 @@
-export { default } from "./Banner";
