+++ src/components/CurriculumComponents/HomePageBanner/index.tsx
@@ -1,0 +1,1 @@
+export { default } from "./HomePageBanner";
