+++ src/pages/ai/index.tsx
@@ -1,12 +1,11 @@
 import { GetStaticProps, NextPage } from "next";
 
-import { isCycleTwoEnabled } from "@/utils/curriculum/features";
 import { DEFAULT_SEO_PROPS } from "@/browser-lib/seo/Seo";
 import AppLayout from "@/components/SharedComponents/AppLayout";
 import getPageProps from "@/node-lib/getPageProps";
 import HomePageTabImageNav from "@/components/GenericPagesComponents/HomePageTabImageNav";
 import { HomePage } from "@/common-lib/cms-types";
-import HomePageBanner from "@/components/CurriculumComponents/Banner";
+import Banners from "@/components/CurriculumComponents/Banners";
 import { HomePageLowerView } from "@/components/GenericPagesViews/HomePageLower/HomePageLower.view";
 import AiTab from "@/components/GenericPagesComponents/AiTab";
 import {
@@ -19,18 +18,9 @@ export type PupilHomePageProps = {
   posts: SerializedPost[];
 };
 
-const cycleTwoEnabled = isCycleTwoEnabled();
-
 const Ai: NextPage<PupilHomePageProps> = (props) => (
   <AppLayout seoProps={DEFAULT_SEO_PROPS} $background={"white"}>
-    {cycleTwoEnabled ? (
-      <HomePageBanner
-        background="lemon"
-        newText="Subjects added"
-        ctaText="See curriculum plans"
-        page="curriculum-landing-page"
-      />
-    ) : null}
+    <Banners />
     <HomePageTabImageNav current={"ai"} />
     <AiTab aria-current="page" />
     <HomePageLowerView posts={props.posts} />
