+++ src/pages/curriculum/index.tsx
@@ -1,11 +1,10 @@
 import { GetStaticProps, NextPage } from "next";
 
-import { isCycleTwoEnabled } from "@/utils/curriculum/features";
 import { DEFAULT_SEO_PROPS } from "@/browser-lib/seo/Seo";
 import AppLayout from "@/components/SharedComponents/AppLayout";
 import getPageProps from "@/node-lib/getPageProps";
 import HomePageTabImageNav from "@/components/GenericPagesComponents/HomePageTabImageNav";
-import HomePageBanner from "@/components/CurriculumComponents/Banner";
+import Banners from "@/components/CurriculumComponents/Banners";
 import { HomePageLowerView } from "@/components/GenericPagesViews/HomePageLower/HomePageLower.view";
 import CurriculumTab from "@/components/GenericPagesComponents/CurriculumTab";
 import {
@@ -13,18 +12,9 @@ import {
   HomePageProps,
 } from "@/pages-helpers/home/getBlogPosts";
 
-const cycleTwoEnabled = isCycleTwoEnabled();
-
 const Curriculum: NextPage<HomePageProps> = (props) => (
   <AppLayout seoProps={DEFAULT_SEO_PROPS} $background={"white"}>
-    {cycleTwoEnabled ? (
-      <HomePageBanner
-        background="lemon"
-        newText="Subjects added"
-        ctaText="See curriculum plans"
-        page="curriculum-landing-page"
-      />
-    ) : null}
+    <Banners />
     <HomePageTabImageNav current={"curriculum"} />
     <CurriculumTab aria-current="page" />
     <HomePageLowerView posts={props.posts} />
